Add Dockerfile
This change adds a Dockerfile which builds the binary in an Alpine linux
image and copies it into a small Alpine image.

It requires an environment variable of HOSTNAME, and takes an optional
variable of WORKGROUP (and defaults to "WORKGROUP").

This can be tested with:
```
docker build -t ${USER}/wsdd2
docker run -ti -e HOSTNAME=MYSERVER ${USER}/wsdd2
```

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>