set LLMNR RR TTL to 30 seconds
RFC 4795

2.8. RR TTL

The responder should insert a pre-configured TTL value in the records
returned in an LLMNR response. A default value of 30 seconds is RECOMMENDED.
In highly dynamic environments (such as mobile ad-hoc networks), the TTL
value may need to be reduced.

Due to the TTL minimalization necessary when caching an RRset, all TTLs in
an RRset MUST be set to the same value.